успешная оптимизация таблички! для получения случайного токена на основе предыдущего мы вместо сохранения хэшмапы с хэшмапами сохраняем хэшмапу с рандомными коллекциями (в случае если табличка только для чтения) в первое время будет только для чтения - так что гуд!

+++ src/main/java/my/lexonix/wordgen/utility/RandomCollection.java
@@ -13,7 +13,7 @@ public class RandomCollection<E> {
         this.random = random;
     }
 
-    public void add(int weight, E result) {
+    public void add(long weight, E result) {
         if (weight <= 0) return;
         total += weight;
         map.put(total, result);
